+++ src/Model/BangDiem.java
@@ -3,29 +3,29 @@ package Model;
 public class BangDiem {
 	public int STT;
 	public String MSSV,Hoten;
-	public int DiemGK,DiemCK,Diemkhac,Tong ;
+	public float DiemGK,DiemCK,Diemkhac,Tong ;
 	public int getSTT() {
 		return STT;
 	}
-	public int getDiemGK() {
+	public float getDiemGK() {
 		return DiemGK;
 	}
 	public void setDiemGK(int diemGK) {
 		DiemGK = diemGK;
 	}
-	public int getDiemCK() {
+	public float getDiemCK() {
 		return DiemCK;
 	}
 	public void setDiemCK(int diemCK) {
 		DiemCK = diemCK;
 	}
-	public int getDiemkhac() {
+	public float getDiemkhac() {
 		return Diemkhac;
 	}
 	public void setDiemkhac(int diemkhac) {
 		Diemkhac = diemkhac;
 	}
-	public int getTong() {
+	public float getTong() {
 		return Tong;
 	}
 	public void setTong(int tong) {
@@ -46,7 +46,7 @@ public class BangDiem {
 	public void setSTT(int sTT) {
 		STT = sTT;
 	}
-	public BangDiem(int sTT, int diemGK, int diemCK, int diemkhac, int tong, String mSSV, String hoten) {
+	public BangDiem(int sTT, float diemGK, float diemCK, float diemkhac, float tong, String mSSV, String hoten) {
 		super();
 		STT = sTT;
 		DiemGK = diemGK;
